requirements.txt
amqp==5.2.0
annotated-types==0.7.0
asgiref==3.8.1
billiard==4.2.0
cachetools==5.5.0
celery==5.4.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==5.1
django-celery-results==2.5.1
django-cors-headers==4.4.0
django-debug-toolbar==4.4.6
django-environ==0.11.2
django-extensions==3.2.3
django-filter==24.3
django-sendgrid-v5==1.2.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
environ==1.0
google-ai-generativelanguage==0.6.6
google-api-core==2.19.2
google-api-python-client==2.144.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-generativeai==0.7.2
googleapis-common-protos==1.65.0
grpcio==1.66.1
grpcio-status==1.62.3
httplib2==0.22.0
idna==3.8
kombu==5.4.0
pillow==10.4.0
prompt_toolkit==3.0.47
proto-plus==1.24.0
protobuf==4.25.4
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.9.1
pydantic_core==2.23.3
pydot==3.0.1
pygraphviz==1.13
PyJWT==2.9.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-http-client==3.3.7
redis==5.0.8
requests==2.32.3
rsa==4.9
sendgrid==6.11.0
six==1.16.0
sqlparse==0.5.1
starkbank-ecdsa==2.2.0
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13

